+++ backend/src/main/java/com/commitmate/re_cord/domain/post/post/controller/ApiV1PostReadController.java
@@ -70,7 +70,7 @@ public class ApiV1PostReadController {
     }
 
 
-    // ✅ 작성자의 다른 게시글 조회
+    // 작성자의 다른 게시글 조회
     @GetMapping("/{userid}/other-posts")
     public List<PostResponseDto> getOtherPostsBySameUser(
             @PathVariable Long userid,
@@ -78,12 +78,22 @@ public class ApiV1PostReadController {
         return postService.getOtherPostsBySameUser(userid, excludePostId);
     }
 
-    // ✅ 작성자의 모든 게시글 조회
+    // 작성자의 모든 게시글 조회
     @GetMapping("/{userid}/posts")
     public List<PostResponseDto> getAllPostsBySameUser(@PathVariable Long userid) {
         return postService.getAllPostsByUser(userid);
     }
 
+    // 작성자의 모든 게시글 리스트
+    @GetMapping("/{userId}/posts/list")
+    public ResponseEntity<Page<PostResponseDto>> getAllPostsBySameUser(
+            @PathVariable Long userId,
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "10") int size
+    ) {
+        Page<PostResponseDto> posts = postService.getAllPostsByUser(userId, page, size);
+        return ResponseEntity.ok(posts);
+    }
 
     // 특정 사용자의 게시글 조회수 총합
     @GetMapping("/views/{userId}")
